Rocket, enemy and health.
Added an enemy behavior, health system for enemies and some rocket logic. TODO: Implement explosion, fine tune speeds, work on the player script.
NOTE: If you guys plan on having another enemy that needs to take damage, make sure that it implements the IDamageable interface.
