+++ Assets/Scripts/AimAtPlayer.cs
@@ -1,5 +1,4 @@
-﻿
-using UnityEngine;
+﻿using UnityEngine;
 
 public class AimAtPlayer : MonoBehaviour
 {
