+++ Assets/Scripts/Constants.cs
@@ -1,0 +1,20 @@
+﻿//Put magic numbers and strings here.
+
+namespace Isopods.Constants
+{
+    public static class PLAYER_CONST
+    {
+        public const string PLAYER_TAG = "Player";
+    }
+
+    public static class ENEMY_CONST
+    {
+        public const float DEFAULT_ENEMY_HEALTH = 50.0f;
+        public const string ENEMY_TAG = "Enemy";
+    }
+
+    public static class WEAPON_CONST
+    {
+        public const float DEFAULT_ROCKET_DAMAGE = 25.0f;
+    }
+}
