+++ Assets/Scripts/Constants.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: e2f248a3c2357d8409ea5c3f55ba2962
+timeCreated: 1525888303
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
