+++ Assets/Scripts/Enemy.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: cdaf33eae8b96c44bbe76e0b7ba4f4b0
+folderAsset: yes
+timeCreated: 1526865519
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
