+++ Assets/Scripts/Enemy/EnemyBehavior.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 09c0d592671da004a829dffe2d41a4ec
+timeCreated: 1526865593
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
