+++ Assets/Scripts/Interfaces.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: e489de728b0b3a640be50319d5c6fb1e
+folderAsset: yes
+timeCreated: 1526866122
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
