+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,0 +1,10 @@
+﻿namespace Isopods.Interfaces
+{
+    public interface IDamageable
+    {
+        float health { get; set; }
+
+        void TakeDamage(float damageTaken);
+    }
+}
+
