+++ Assets/Scripts/Interfaces/IDamageable.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: ef91dbbd4d882f64394cba257e61c31e
+timeCreated: 1526866139
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
