--- Assets/Scripts/Jump.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 197ea3e05995c2240b1686b85b714227
-timeCreated: 1524985613
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
