--- Assets/Scripts/NewFollow.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 655604d1a523b404b814ef5fabaf4117
-timeCreated: 1524941913
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
