+++ Assets/Scripts/Player.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 6c27f2be46832ca45b27aeb1f16a672a
+folderAsset: yes
+timeCreated: 1526427825
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
