+++ Assets/Scripts/Player/Move.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: b03d915456ccf0b409f78014339e9c98
+timeCreated: 1524717718
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
