+++ Assets/Scripts/Weapons.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 151cfa55c4e566e4491f7427dab4cdf6
+folderAsset: yes
+timeCreated: 1525888322
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
