+++ Assets/Scripts/Weapons/Bazooka.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 090abfc99a6083241ab8d2408f33e068
+timeCreated: 1525888348
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
