--- Assets/Scripts/projectile.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: cf16661af463800489871a7d8b0638fc
-timeCreated: 1525474509
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
