+++ Assets/prefabs/Rocket.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 64300e6b29490894f9e410886ac4b512
+timeCreated: 1526258033
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
