+++ ProjectSettings/TagManager.asset
@@ -5,6 +5,7 @@ TagManager:
   serializedVersion: 2
   tags:
   - ground
+  - Enemy
   layers:
   - Default
   - TransparentFX
